+++ src/main/java/com/github/yeetmanlord/somanyenchants/common/blocks/smelters/smoker/EnchantedSmokerBlock.java
@@ -56,7 +56,7 @@ public class EnchantedSmokerBlock extends AbstractEnchantedSmelterBlock
 		if (stateIn.getValue(LIT))
 		{
 			double d0 = (double) pos.getX() + 0.5D;
-			double d1 = (double) pos.getY();
+			double d1 = pos.getY();
 			double d2 = (double) pos.getZ() + 0.5D;
 
 			if (rand.nextDouble() < 0.1D)
